Rust: rename daskr module aux -> auxiliary (Windows reserved name)
`aux` is a reserved device name on Windows (like con, prn, nul), so
git cannot check out `daskr/src/aux.rs` on Windows. This breaks the
merge/checkout step for every PR built against master since the
Rust-port PR #15859 introduced the file.

Rename the module and its file to `auxiliary` and update all
references (daskr solver/tests, wasm_jit sim_driver, wasm_jit_runtime
standalone) plus the intra-doc link in lib.rs.
